immich: bump to 2.3.1
2.2.3 saw the introduction of a version check bug, which may have
evaded many users if they've been using latest/stable (as that's only
had 2.1.x releases). 2.3.1 resolves this issue, but introduces some
new immich-core plugin (which is still being investigated, the current
measure is merely a stop-gap to allow Immich to run as before), and
libGL is now a hard-requirement for the underlying python in the ML
implementation (which we provide, instead of jellyfin-ffmpeg).

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>