Makefile: major update
It should actually function now!

Up-shots:
* All inter-snap dependencies resolve appropriately, with none
  being needed at install-time (for the purposes of cross-compiling,
  as installing the dependency... *should* be impossible... Although on
  second thought it may not be...).
* All snaps build, and can be built independently (dependencies aside)
* All snaps can be packed, and can be packed independently (dependencies aside)
* All snaps can be released to lastest/edge!

Next phase is to implement promotion up the chain
(edge->beta->candidate->stable), but I need to come up with a solid
heuristic and cadence for doing this.

After that, it's just a matter of creating a cronjob to execute this
Makefile every so often. Perhaps on a trigger (new version, USN), but at
least every week or so...

Signed-off-by: Dilyn Corner <dilyn@dilyn.cc>